Added new shortcuts for table view to move cursor
