+++ src/shortcuts/moveCursorBackward.js
@@ -1,0 +1,37 @@
+import { store } from '../store'
+
+// util
+import {
+  contextOf,
+  isDocumentEditable,
+  pathToContext,
+} from '../util'
+
+// action-creators
+import { cursorBack } from '../action-creators/cursorBack'
+import { outdent } from '../action-creators/outdent'
+
+// selectors
+import pathToThoughtsRanked from '../selectors/pathToThoughtsRanked'
+import attributeEquals from '../selectors/attributeEquals'
+
+export default {
+  id: 'moveCursorBackward',
+  name: 'Move Cursor Backward',
+  description: `Move the current thought to the next sibling of its context or to previous column in table view.`,
+  keyboard: { key: 'Tab', shift: true },
+  canExecute: () => isDocumentEditable() && store.getState().cursor,
+  exec: e => {
+    const state = store.getState()
+    const { cursor } = state
+
+    if (cursor.length < 2) return
+
+    const thoughtsRanked = pathToThoughtsRanked(state, cursor)
+    // contextOf twice because we are checking if this thought is in column 2 of a table
+    const contextGrandparent = contextOf(contextOf(pathToContext(thoughtsRanked)))
+    const isTable = attributeEquals(state, contextGrandparent, '=view', 'Table')
+
+    store.dispatch(isTable ? cursorBack({ target: e.target }) : outdent())
+  }
+}
